pyproject.toml
[tool.poetry]
name = "djongo-tutorial"
version = "0.1.0"
description = ""
authors = ["Jc Din <ozknightwalker62@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
djongo = "^1.3.6"
Django = "^3.2.6"
drf-yasg = "^1.20.0"
django-hashid-field = "^3.3.0"
django-redis = "^5.0.0"
channels = "^3.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
